refactor: reorganize Python code into maintainable structure
Consolidate scattered Python files into a logical, organized structure
while preserving all existing functionality.

Changes:
- Create shared library package (automation/test-execution/shared/)
  - Extract common utilities: io_utils.py (JSON I/O, time formatting)
  - Add vllm_metrics.py placeholder for future consolidation

- Reorganize scripts by purpose:
  - ansible/ → Scripts invoked by Ansible playbooks
  - conversion/ → Result conversion tools

- Update all references:
  - Ansible playbooks: llm-benchmark.yml, llm-benchmark-auto.yml, log-to-mlflow.yml
  - Bash scripts: extract-all-timings.sh

- Improve test infrastructure:
  - Add root conftest.py for pytest path configuration
  - Add __init__.py to create proper Python packages
  - Maintain sys.path imports for test compatibility

File moves (git tracked):
- ansible/scripts/*.py → scripts/ansible/*.py
- results/scripts/*.py → scripts/conversion/*.py
- scripts/python/*.py → scripts/conversion/*.py

Verification:
✅ All 52 unit tests passing
✅ Scripts functional
✅ Ansible playbooks valid
✅ No breaking changes
