+++ automation/test-execution/ansible/filter_plugins/__init__.py
@@ -1,0 +1,1 @@
+"""Ansible Jinja2 filter plugins for CPU topology and NUMA operations."""
