+++ automation/test-execution/scripts/ansible/extract_benchmark_timings.py
@@ -6,14 +6,16 @@ for each benchmark run, including duration, requests processed, and temporal met
 The extracted data is appended to the test-metadata.json file for later analysis.
 """
 
-import json
 import sys
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Tuple
 
-# Time conversion constants
-SECONDS_PER_HOUR = 3600
-SECONDS_PER_MINUTE = 60
+# Add shared library to path
+_script_dir = Path(__file__).parent
+_shared_dir = _script_dir.parent.parent / "shared"
+sys.path.insert(0, str(_shared_dir))
+
+from io_utils import load_json_file, save_json_file, format_duration  # noqa: E402
 
 
 def extract_timings(bench_data: Dict[str, Any]) -> Tuple[List[Dict[str, Any]], float]:
@@ -47,50 +49,6 @@ def extract_timings(bench_data: Dict[str, Any]) -> Tuple[List[Dict[str, Any]], f
         benchmark_timings.append(timing)
 
     return benchmark_timings, total_duration
-
-
-def format_duration(total_seconds: float) -> str:
-    """Format duration as HH:MM:SS string.
-
-    Args:
-        total_seconds: Duration in seconds
-
-    Returns:
-        Formatted string in HH:MM:SS format
-    """
-    hours = int(total_seconds // SECONDS_PER_HOUR)
-    minutes = int((total_seconds % SECONDS_PER_HOUR) // SECONDS_PER_MINUTE)
-    seconds = int(total_seconds % SECONDS_PER_MINUTE)
-    return f"{hours}:{minutes:02d}:{seconds:02d}"
-
-
-def load_json_file(file_path: Path) -> Dict[str, Any]:
-    """Load and parse a JSON file.
-
-    Args:
-        file_path: Path to the JSON file
-
-    Returns:
-        Parsed JSON data as a dictionary
-
-    Raises:
-        FileNotFoundError: If the file doesn't exist
-        json.JSONDecodeError: If the file contains invalid JSON
-    """
-    with open(file_path, 'r', encoding='utf-8') as f:
-        return json.load(f)
-
-
-def save_json_file(file_path: Path, data: Dict[str, Any]) -> None:
-    """Save data to a JSON file with pretty formatting.
-
-    Args:
-        file_path: Path to the output JSON file
-        data: Dictionary to save as JSON
-    """
-    with open(file_path, 'w', encoding='utf-8') as f:
-        json.dump(data, f, indent=2)
-        f.write('\n')  # Add trailing newline
 
 
 def main() -> int:
@@ -131,10 +89,6 @@ def main() -> int:
 
     except FileNotFoundError as e:
         print(f"Warning: File not found: {e.filename}", file=sys.stderr)
-        return 0  # Don't fail the playbook - this is a non-critical enhancement
-
-    except json.JSONDecodeError as e:
-        print(f"Warning: Invalid JSON in file: {e}", file=sys.stderr)
         return 0  # Don't fail the playbook - this is a non-critical enhancement
 
     except KeyError as e:
