+++ automation/test-execution/shared/__init__.py
@@ -1,0 +1,9 @@
+"""
+Shared utilities for vLLM CPU Performance Evaluation project.
+
+This package contains common utilities used across multiple scripts and tools:
+- io_utils: JSON file I/O and common file operations
+- vllm_metrics: vLLM Prometheus metrics parsing helpers
+"""
+
+__version__ = "0.1.0"
