+++ automation/test-execution/shared/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for shared utilities."""
