+++ automation/test-execution/shared/vllm_metrics.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+"""vLLM Prometheus metrics parsing utilities.
+
+This module provides helpers for parsing vLLM server metrics from Prometheus format.
+Currently a placeholder for future consolidation of metrics parsing logic.
+"""
+
+# Placeholder for future consolidation of vLLM metrics parsing
+# Currently log_to_mlflow.py and convert_single.py have different implementations
+# that serve different purposes. Common helpers can be added here as patterns emerge.
